pyproject.toml
[tool.poetry]
name = "advantis"
version = "0.1.0"
description = "Agentic solutions for Advantis"
authors = ["Nuwan Waidyanatha <farmraider@protonmail>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
agentops = { version = "^0.3.0", optional = true }
apache-beam="^2.59.0"
appdirs = "^1.4.4"
arrow = "^1.2.3"
auth0-python = "^4.7.1"
beautifulsoup4 = "^4.11.1"
boto3 = "^1.37.4"
#botocore = "^1.27.89"
chromadb = "<0.5.0"
crewai = "<0.108.0"
crewai-tools = { version = "^0.12.1", optional = true }
click = "^8.1.7"
db-dtypes = "^1.3.1"
duckduckgo_search = "^6.3.0"
embedchain = "^0.1.114"
findspark = "^2.0.1"
flask = "^2.2.2"
geopandas = "^0.13.2"
#google-auth = "^2.12.0"
#google-auth-oauthlib = "^0.5.0"
#google-pasta = "^0.2.0"
#google-cloud-bigquery = "^3.27.0"
#google-cloud-bigquery-storage = "^2.27.0"
#google-cloud-storage = "^2.10.0"
#graphene = "^3.1"
#graphql-core = "^3.2.1"
ipykernel = "^6.15.1"
ipython = "^8.4.0"
isodate = "^0.6.1"
jsonref = "^1.1.0"
json-repair = "^0.25.2"
jupyter = "^1.0.0"
jupyter-client = "^7.4.4"
langchain = "0.2.16"
langchain-core = "0.2.38"
langchain-chroma = "^0.1.4"
langchain-groq = "^0.1.9"
litellm = "^1.44.22"
#networkx = "^2.8.6"
nltk = "^3.8.1"
notebook = "^7.0.0"
numpy = "^1.23.2"
numba = "<=0.58.1"
oauthlib = "^3.2.2"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
pandas = "^1.4.3"
pillow = "^9.2.0"
#pymongo="4.3.3"
#poetry = "^1.8.3"
pgvector = "^0.3.5"
#pycaret = "<4.0.0"
pydantic = "^2.4.2"
pyarrow = ">=3.0.0,<17.0.0"
pyspark = "3.4.1"
python-dotenv = "^1.0.0"
psycopg-binary = "^3.1.1"
psycopg2-binary = "^2.9.6"
pysqlite3 = "^0.5.4"
#requests = "^2.28.0"
schema = "^0.7.7"
scikit-learn = ">1.4.0"
scipy = "^1.9.1"
spacy = "^3.6.1"
sqlalchemy = ">=2.0.27,<3.0.0"
#tensorflow = "^2.18.0"
#tensorflow-probability = "^0.25.0"
#tf-keras = "^2.18.0"
torch = "^1.12.1"
tqdm = "^4.64.1"
transformers = "^4.22.2"
wheel = "^0.38.4"
#xeus-python = "^0.15.12"
ydata-profiling = "^4.12.1"
google-auth = "^2.38.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.166.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
